updating works as well no js at all and update is a little bit lumbersexual

+++ src/main/java/com/codecool/spring/repository/TodoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TodoRepository extends JpaRepository<Todo, Long> {
+public interface TodoRepository extends JpaRepository<Todo,Integer> {
 
     List<Todo> findByStatus(Status status);
 }
